fix-bug: model transmission error
