+++ src/main/java/group1/intern/SellingAndManagingShoesApplication.java
@@ -1,0 +1,12 @@
+package group1.intern;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SellingAndManagingShoesApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(SellingAndManagingShoesApplication.class, args);
+    }
+
+}
