+++ src/main/java/group1/intern/ServletInitializer.java
@@ -1,0 +1,13 @@
+package group1.intern;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(SellingAndManagingShoesApplication.class);
+	}
+
+}
