+++ src/main/java/group1/intern/bean/AccountInfo.java
@@ -1,0 +1,5 @@
+package group1.intern.bean;
+
+public class AccountInfo {
+
+}
