+++ src/main/java/group1/intern/config/SecurityConfig.java
@@ -1,0 +1,7 @@
+package group1.intern.config;
+
+public class SecurityConfig {
+
+}
+
+
