+++ src/main/java/group1/intern/controller/HomeController.java
@@ -1,0 +1,12 @@
+package group1.intern.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+    @GetMapping("/")
+    public String homepage() {
+        return "index";
+    }
+}
