+++ src/main/java/group1/intern/controller/LoginController.java
@@ -1,0 +1,4 @@
+package group1.intern.controller;
+
+public class LoginController {
+}
