+++ src/main/java/group1/intern/controller/RegisterController.java
@@ -1,0 +1,14 @@
+package group1.intern.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class RegisterController {
+
+    @GetMapping("/register")
+    public String showRegisterPage() {
+        return "screens/auth/register";
+    }
+
+}
