+++ src/main/java/group1/intern/model/Account.java
@@ -1,0 +1,45 @@
+package group1.intern.model;
+
+import group1.intern.model.Enum.AccountRole;
+import jakarta.persistence.*;
+import lombok.*;
+import lombok.experimental.SuperBuilder;
+
+import java.time.LocalDate;
+import java.util.List;
+
+
+@Entity
+@Getter
+@Setter
+@SuperBuilder
+@AllArgsConstructor
+@NoArgsConstructor
+@Table(name = "accounts")
+public class Account extends EntityBase {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String email;
+    private String displayName;
+    private String password;
+    @Enumerated(EnumType.STRING)
+    private AccountRole role;
+    private String fullName;
+    private LocalDate dateOfBirth;
+    private String address;
+    private String phoneNumber;
+    private Boolean gender;
+    private String avatarUrl;
+    private Boolean isActivated = true;
+    // One-to-Many relationship with RefreshToken
+    @OneToMany(mappedBy = "account",fetch = FetchType.LAZY)
+    private List<RefreshToken> refreshTokens;
+    // One-to-Many relationship with ShoppingCart
+    @OneToMany(mappedBy = "account",fetch = FetchType.LAZY)
+    private List<ShoppingCart> shoppingCarts;
+    // One-to-Many relationship with Order
+    @OneToMany(mappedBy = "account",fetch = FetchType.LAZY)
+    private List<Order> orders;
+}
+
