+++ src/main/java/group1/intern/model/Embeddables/ProductDescription.java
@@ -1,0 +1,15 @@
+package group1.intern.model.Embeddables;
+
+import lombok.*;
+
+import java.util.List;
+
+@Builder
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ProductDescription {
+    private String text;
+    private List<String> images;
+}
