+++ src/main/java/group1/intern/model/Enum/AccountRole.java
@@ -1,0 +1,5 @@
+package group1.intern.model.Enum;
+
+public enum AccountRole {
+    ADMIN, CUSTOMER, SELLER
+}
