+++ src/main/java/group1/intern/model/Enum/OrderStatus.java
@@ -1,0 +1,5 @@
+package group1.intern.model.Enum;
+
+public enum OrderStatus {
+    WAIT, CONFIRM, REJECT, CANCEL, RECEIVED, DONE
+}
