+++ src/main/java/group1/intern/model/Enum/ProductGender.java
@@ -1,0 +1,5 @@
+package group1.intern.model.Enum;
+
+public enum ProductGender {
+    MALE, FEMALE, UNISEX
+}
