+++ src/main/java/group1/intern/repository/AccountRepository.java
@@ -1,0 +1,4 @@
+package group1.intern.repository;
+
+public class AccountRepository {
+}
