+++ src/main/java/group1/intern/repository/ConstantRepository.java
@@ -1,0 +1,9 @@
+package group1.intern.repository;
+
+import group1.intern.model.Constant;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ConstantRepository extends JpaRepository<Constant, Integer> {
+}
