+++ src/main/java/group1/intern/repository/ProductRepository.java
@@ -1,0 +1,10 @@
+package group1.intern.repository;
+
+import group1.intern.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+
+}
