+++ src/main/java/group1/intern/service/AccountService.java
@@ -1,0 +1,4 @@
+package group1.intern.service;
+
+public class AccountService {
+}
