--- src/main/java/sun/intern/ServletInitializer.java
@@ -1,13 +1,0 @@
-package sun.intern;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(SellingAndManagingShoesApplication.class);
-	}
-
-}
