--- src/main/java/sun/intern/bean/AccountInfo.java
@@ -1,5 +1,0 @@
-package sun.intern.bean;
-
-public class AccountInfo {
-
-}
