--- src/main/java/sun/intern/config/SecurityConfig.java
@@ -1,7 +1,0 @@
-package sun.intern.config;
-
-public class SecurityConfig {
-
-}
-
-
