--- src/main/java/sun/intern/controller/HomeController.java
@@ -1,12 +1,0 @@
-package sun.intern.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-    @GetMapping("/")
-    public String homepage() {
-        return "index";
-    }
-}
