--- src/main/java/sun/intern/controller/LoginController.java
@@ -1,4 +1,0 @@
-package sun.intern.controller;
-
-public class LoginController {
-}
