--- src/main/java/sun/intern/controller/RegisterController.java
@@ -1,14 +1,0 @@
-package sun.intern.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class RegisterController {
-    
-    @GetMapping("/register")
-    public String showRegisterPage() {
-        return "screens/auth/register";
-    }
-
-}
