--- src/main/java/sun/intern/model/Embeddables/ProductDescription.java
@@ -1,15 +1,0 @@
-package sun.intern.model.Embeddables;
-
-import lombok.*;
-
-import java.util.List;
-
-@Builder
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ProductDescription {
-    private String text;
-    private List<String> images;
-}
