--- src/main/java/sun/intern/model/Enum/AccountRole.java
@@ -1,5 +1,0 @@
-package sun.intern.model.Enum;
-
-public enum AccountRole {
-    ADMIN, CUSTOMER, SELLER
-}
