--- src/main/java/sun/intern/model/Enum/OrderStatus.java
@@ -1,5 +1,0 @@
-package sun.intern.model.Enum;
-
-public enum OrderStatus {
-    WAIT, CONFIRM, REJECT, CANCEL, RECEIVED, DONE
-}
