--- src/main/java/sun/intern/model/Enum/ProductGender.java
@@ -1,5 +1,0 @@
-package sun.intern.model.Enum;
-
-public enum ProductGender {
-    MALE, FEMALE, UNISEX
-}
