--- src/main/java/sun/intern/model/Order.java
@@ -1,37 +1,0 @@
-package sun.intern.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-import sun.intern.model.Enum.OrderStatus;
-
-import java.time.LocalDateTime;
-import java.util.List;
-@Entity
-@Getter
-@Setter
-@SuperBuilder
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "orders")
-public class Order extends EntityBase {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private double totalPrice;
-    private String phoneNumber;
-    private String address;
-    @Enumerated(EnumType.STRING)
-    private OrderStatus status;
-
-    @ManyToOne
-    @JoinColumn(name = "account_id", nullable = false)
-    private Account account;
-
-    // One-to-Many relationship with OrderDetail
-    @OneToMany(mappedBy = "order")
-    private List<OrderDetail> orderDetails;
-}
