--- src/main/java/sun/intern/model/Product.java
@@ -1,57 +1,0 @@
-package sun.intern.model;
-
-import io.hypersistence.utils.hibernate.type.json.JsonType;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-
-import org.hibernate.annotations.Type;
-import org.hibernate.annotations.UpdateTimestamp;
-import sun.intern.model.Embeddables.ProductDescription;
-import sun.intern.model.Enum.ProductGender;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Entity
-@Getter
-@Setter
-@SuperBuilder
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "products")
-public class Product extends EntityBase{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private int originPrice;
-    private Integer discount;
-    private String name;
-    @Enumerated(EnumType.STRING)
-    private ProductGender gender;
-    @Type(JsonType.class)
-    @Column(columnDefinition = "json")
-    private ProductDescription description;
-    // One-to-Many relationship with ProductImage
-    @OneToMany(mappedBy = "product")
-    private List<ProductImage> images;
-    private LocalDateTime deletedAt;
-
-    @ManyToOne
-    @JoinColumn(name = "category_id")
-    private Constant category;
-
-    @ManyToOne
-    @JoinColumn(name = "style_id")
-    private Constant style;
-
-    @ManyToOne
-    @JoinColumn(name = "material_id")
-    private Constant material;
-
-    @OneToMany(mappedBy = "product",fetch = FetchType.LAZY)
-    private List<ProductDetail> productDetails;
-}
