--- src/main/java/sun/intern/model/ProductDetail.java
@@ -1,40 +1,0 @@
-package sun.intern.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Entity
-@Getter
-@Setter
-@SuperBuilder
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "product_details")
-public class ProductDetail extends EntityBase{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(insertable=false, updatable=false)
-    private Integer id;
-    private int price;
-    private LocalDateTime deletedAt;
-    private int quantity;
-    @ManyToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    @ManyToOne
-    @JoinColumn(name = "color_id")
-    private Constant color;
-
-    @ManyToOne
-    @JoinColumn(name = "size_id")
-    private Constant size;
-
-}
