--- src/main/java/sun/intern/model/ProductImage.java
@@ -1,25 +1,0 @@
-package sun.intern.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@Setter
-@SuperBuilder
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "product_images")
-public class ProductImage {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String url;
-    @ManyToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
-}
