--- src/main/java/sun/intern/model/RefreshToken.java
@@ -1,27 +1,0 @@
-package sun.intern.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@Setter
-@SuperBuilder
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "refresh_tokens")
-public class RefreshToken extends EntityBase {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    @Column(insertable=false, updatable=false)
-    private String token;
-    @ManyToOne
-    @JoinColumn(name = "account_id")
-    private Account account;
-}
-
