--- src/main/java/sun/intern/repository/AccountRepository.java
@@ -1,4 +1,0 @@
-package sun.intern.repository;
-
-public class AccountRepository {
-}
