--- src/main/java/sun/intern/repository/ProductRepository.java
@@ -1,10 +1,0 @@
-package sun.intern.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import sun.intern.model.Product;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-
-}
