--- src/main/java/sun/intern/service/AccountService.java
@@ -1,4 +1,0 @@
-package sun.intern.service;
-
-public class AccountService {
-}
